+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -1,7 +1,6 @@
-package roomescape;
+package roomescape.domain;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.function.Supplier;
 import java.util.stream.Stream;
 import org.assertj.core.api.Assertions;
@@ -10,7 +9,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import roomescape.domain.Reservation;
 
 class ReservationTest {
     @Test
