+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -1,9 +1,8 @@
-package roomescape;
+package roomescape.domain;
 
 import java.util.function.Supplier;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.ReservationTime;
 
 class ReservationTimeTest {
     @Test
