+++ src/test/java/roomescape/domain/ReservationsTest.java
@@ -1,13 +1,10 @@
-package roomescape;
+package roomescape.domain;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
 import java.time.LocalDate;
-import java.time.LocalTime;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.Reservation;
-import roomescape.domain.Reservations;
 
 class ReservationsTest {
     @Test
